今日工作

+++ src/main/java/com/aokai/parking/model/vo/TodayOrderResp.java
@@ -31,5 +31,5 @@ public class TodayOrderResp implements Serializable {
     /**
      * 今日总营业额
      */
-    private Integer costTotal;
+    private Double costTotal;
 }
